# src/swagger/swagger.yml
openapi: 3.0.0
paths:
...
  /auth/employee/login:
    post:
      operationId: AuthController_employeeLogin
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/loginEmployeeDto'
      responses:
        '201':
          description: ''
      tags: &ref_10
        - employee
  /auth/employee/register:
    post:
      operationId: AuthController_employeeRegister
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/registerEmployeeDto'
      responses:
        '201':
          description: ''
      tags: *ref_10
  /auth/employee/{id}/details:
    get:
      operationId: AuthController_getEmployeeDetails
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: number
      responses:
        '200':
          description: Gives the order details by given id
          content:
            application/json:
              examples:
                success:
                  description: Gives the menu details
                  value:
                    succes: true
                    data:
                      menuitem_id: 2
                      menu_itemname: manchuria
                      price: 10
                      tax: 5
                      date:
                        createdDate: '2023-08-17T08:31:36.802Z'
                        updatedDate: '2023-08-24T04:40:27.390Z'
        '400':
          description: Occurs when given Id is invalid
          content:
            application/json:
              examples:
                InvalidId:
                  description: Occurs when given Id is invalid
                  value:
                    message: Given_id_is_not_found
      tags: *ref_10
  /auth/employee/roles:
    post:
      operationId: AuthController_createRoles
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/createRoleDto'
      responses:
        '200':
          description: creates a role
          content:
            application/json:
              examples:
                success:
                  description: Gives the role deatils that have been posted
                  value:
                    succes: true
                    data:
                      name: Customer
                      description: customer has permission to order
                      id: 5
        '409':
          description: Occurs when you are trying to post the same role again
          content:
            application/json:
              examples:
                conflict:
                  description: Same Role posted again
                  value:
                    error: Conflict Error
                    message: 'Given_role_is_already_present '
      tags: *ref_10
  /auth/employee/role/{id}:
    put:
      operationId: AuthController_updateRoles
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/updateRoleDto'
      responses:
        '200':
          description: Role is updated succefully
          content:
            application/json:
              examples:
                success:
                  description: Role is updated with given name in database
                  value:
                    succes: true
                    data:
                      id: 1
                      employee_Id: A1
                      employee_Name: prem
                      status: Active
                      email: prem3a3
                      phoneNumber: 8367
                      date:
                        createdDate: '2023-08-22T05:30:40.783Z'
                        updatedDate: '2023-08-22T09:38:24.000Z'
                      roles:
                        id: 1
                        name: Manager
        '400':
          description: Occurs when given path varaible is invalid
          content:
            application/json:
              examples:
                IdNotFound:
                  description: Role is updated with given name in database
                  value:
                    message: Id not found to update the role
      tags: *ref_10
info:
  title: chandu
  description: Your API Description
  version: '1.0'
  contact: {}
...
servers:
  - url: http://localhost:3000
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
  schemas:
...
    loginEmployeeDto:
      type: object
      properties: {}
    registerEmployeeDto:
      type: object
      properties: {}
    createRoleDto:
      type: object
      properties: {}
    updateRoleDto:
      type: object
      properties:
        name:
          type: string
          example: Manager
      required:
        - name